pyproject.toml
[tool.poetry]
name = "nshm-toshi-api"
version = "0.2.0"
description = "the object store used by NZSHM project"
authors = ["Chris Chamberlain <chrisbc@artisan.co.nz>"]
license = "AGPL3"
readme = "README.md"
packages = [{include = "graphql_api"}]

[tool.poetry.scripts]
elastic_cli = 'scripts.elastic_cli:main'

[tool.poetry.dependencies]
python = "^3.10"

flask = "^3.0.3"
flask-cors = "^4.0.1"
graphql-server = {extras = ["flask"], version = "^3.0.0b7"}
graphene = "^3.3"

boto3 = "^1.26.98"
botocore = "^1.29.98"

nzshm-common = "^0.4.0"
backoff = "2.2.1"
pynamodb = "^5.5.0"
python-benedict = "^0.33.0"
requests-aws4auth = "^1.2.3"
pyyaml = "^6.0"
elasticsearch-dsl = ">=7.0.0,<8.0"
elasticsearch = ">=7.0.0,<7.14"
gql = "^3.5.0"
requests-toolbelt = "^1.0.0"

[tool.poetry.group.dev.dependencies]
click = "^8.1.7"
tox = "^4.2.8"
black  = { version = "^24.3.0"}
isort  = { version = "^5.12.0"}
flake8  = "^7.0.0"
flake8-docstrings = { version = "^1.6.0", optional = true }
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
moto = "^4.2.7"
requests-mock = "^1.11.0"
virtualenv  = { version = "^20.2.2", optional = true}
pip  = { version = "^23.2.1", optional = true}
bump2version = {version = "^1.0.1"}
mypy = "^1.5.1"
types-pyyaml = "^6.0.12.20240311"
autoflake = "^2.3.1"
types-python-dateutil = "^2.9.0.20240316"
types-requests = "^2.32.0.20240602"


[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
